Sace Memory and Cpu usage

+++ src/main/java/com/example/fedex_backend/models/program/ProgramDTO.java
@@ -12,5 +12,5 @@ import lombok.Setter;
 public class ProgramDTO {
 
   private String name;
-  private Long memory;
+  private Double memory;
 }
